+++ tests/src/all_suites.adb
@@ -5,6 +5,7 @@ with Context_Tests;
 with Engine_Tests;
 with Standard_Tests;
 with Composition_Tests;
+with Collection_Tests;
 
 package body All_Suites is
 
@@ -19,6 +20,7 @@ package body All_Suites is
       Engine_Tests.Add_Tests (Result);
       Standard_Tests.Add_Tests (Result);
       Composition_Tests.Add_Tests (Result);
+      Collection_Tests.Add_Tests (Result);
       return Result;
    end Suite;